src/yolo/requirement.txt
absl-py==0.7.1
astor==0.8.0
bleach==1.5.0
cycler==0.10.0
gast==0.2.2
grpcio==1.21.1
h5py==2.9.0
html5lib==0.9999999
Keras==2.1.5
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.0.3
numpy==1.16.4
Pillow==6.0.0
protobuf==3.8.0
pyparsing==2.4.0
python-dateutil==2.8.0
PyYAML==5.1.1
scipy==1.3.0
six==1.12.0
tensorboard==1.6.0
tensorflow==1.14.0
termcolor==1.1.0
Werkzeug==0.15.4